Changed doubles into floats

+++ spring-boot/src/main/java/com/mattoi/frambo_mock/product/Product.java
@@ -1,12 +1,12 @@
 package com.mattoi.frambo_mock.product;
 
 public record Product(
-        Integer id,
-        String name,
-        String description,
-        String photoUrl,
-        Integer netWeight,
-        Float price,
-        Boolean inStock,
-        String category) {
+                Integer id,
+                String name,
+                String description,
+                String photoUrl,
+                Integer netWeight,
+                Double price,
+                Boolean inStock,
+                String category) {
 }
